Use cookies and add authentication to uploads and backups
